+++ Anvil/Camera.cpp
@@ -1,6 +1,6 @@
 #include "Camera.h"
 
-#include <Sprocket/Core/Events.h>
+#include <Sprocket/Core/events.h>
 #include <Sprocket/Core/Window.h>
 #include <Sprocket/Graphics/Viewport.h>
 #include <Sprocket/Core/input_codes.h>
