+++ Game/Main.m.cpp
@@ -1,7 +1,7 @@
 #include <Game/Game.h>
 
 #include <Sprocket/Core/Window.h>
-#include <Sprocket/Core/GameLoop.h>
+#include <Sprocket/Core/game_loop.h>
 
 using namespace spkt;
 
@@ -9,7 +9,7 @@ int main()
 {
     spkt::window window("Game");
     Game game(&window);
-    spkt::RunOptions options;
-    options.showFramerate = true;
-    return spkt::Run(game, window, options);
+    spkt::run_options options;
+    options.show_frame_rate = true;
+    return spkt::run(game, window, options);
 }
