+++ Sprocket/Scene/Systems/input_system.cpp
@@ -1,6 +1,6 @@
 #include "input_system.h"
 
-#include <Sprocket/Core/Events.h>
+#include <Sprocket/Core/events.h>
 #include <Sprocket/Scene/ecs.h>
 #include <Sprocket/Scene/Scene.h>
