+++ Sprocket/UI/SimpleUI.cpp
@@ -1,7 +1,7 @@
 #include "SimpleUI.h"
 
 #include <Sprocket/Core/Window.h>
-#include <Sprocket/Core/Events.h>
+#include <Sprocket/Core/events.h>
 #include <Sprocket/Core/input_codes.h>
 
 #include <fmt/core.h>
